+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Accept remote identity proofs created using `eddsa-jcs-2022` cryptosuite.
+
 ## [3.20.0] - 2025-03-15
 
 ### Added